fix(platform): enlarge image previews and add click-to-zoom in composer
- Increase sent message image thumbnails from h-7/56px to h-14/120px
- Increase composer attachment image chips from h-6/w-6 to h-14/120px
- Add click-to-zoom lightbox to composer image previews
- Extract ImageLightbox into shared zero-attachment-chips module

Closes #5008
